Machine learning
Designing good models can be tough, e.g. for postal addresses, where models need to be updated to cater for newly discovered examples that don't fit with the existing model. Machine learning from examples and counter examples can help to reduce the burden. Not only for the data models, but also for the rules that operate on them.